pom changes (i *think* this is how it should look). Removed SynonymFilter from SimpleAnalyzer.
