+++ java/pom.xml
@@ -1,4 +1,5 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -11,6 +12,10 @@
         <version>7</version>
     </parent>
 
+    <artifactId>hebmorph-lucene</artifactId>
+    <version>2.0-SNAPSHOT</version>
+    <packaging>pom</packaging>
+
     <build>
         <plugins>
             <plugin>
@@ -19,23 +24,21 @@
                     <source>1.7</source>
                     <target>1.7</target>
                 </configuration>
+                <version>3.2</version>
             </plugin>
-	<plugin>
-	    <groupId>org.apache.maven.plugins</groupId>
-	    <artifactId>maven-release-plugin</artifactId>
-	    <version>2.2.1</version>
-	    <executions>
-	        <execution>
-	            <id>default</id>
-	            <goals>
-	                <goal>perform</goal>
-	            </goals>
-	            <configuration>
-	                <pomFileName>java/pom.xml</pomFileName>
-	            </configuration>
-	        </execution>
-	    </executions>
-	</plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-release-plugin</artifactId>
+                <version>2.2.1</version>
+                <executions>
+                    <execution>
+                        <id>default</id>
+                        <goals>
+                            <goal>perform</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
         </plugins>
     </build>
 
@@ -47,33 +50,7 @@
         </repository>
     </repositories>
 
-    <groupId>com.code972.hebmorph</groupId>
-    <artifactId>hebmorph-lucene</artifactId>
-    <packaging>jar</packaging>
-    <version>2.0.0-SNAPSHOT</version>
-    <name>hebmorph-lucene</name>
-    <description>Enabling Hebrew search through Lucene. Part of the HebMorph project, for making Hebrew properly searchable</description>
-    <url>http://code972.com/</url>
-    <licenses>
-        <license>
-            <name>Affero GPL3</name>
-            <url>http://www.gnu.org/licenses/agpl.html</url>
-            <distribution>repo</distribution>
-            <comments>The GNU Affero General Public License is a free, copyleft license for software and other kinds of works, specifically designed to ensure cooperation with the community in the case of network server software.</comments>
-        </license>
-    </licenses>
-    <developers>
-        <developer>
-            <id>synhershko</id>
-            <name>Itamar Syn-Hershko</name>
-            <url>http://code972.com</url>
-        </developer>
-    </developers>
-    <scm>
-        <url>https://github.com/synhershko/HebMorph</url>
-	<connection>scm:git:ssh://git@github.com/synhershko/HebMorph.git</connection>
-	<developerConnection>scm:git:ssh://git@github.com/synhershko/HebMorph.git</developerConnection>
-    </scm>
+
     <dependencies>
         <dependency>
             <groupId>org.apache.lucene</groupId>
@@ -130,9 +107,9 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-        	<groupId>org.apache.lucene</groupId>
-        	<artifactId>lucene-queryparser</artifactId>
-        	<version>${lucene.version}</version>
+            <groupId>org.apache.lucene</groupId>
+            <artifactId>lucene-queryparser</artifactId>
+            <version>${lucene.version}</version>
         </dependency>
     </dependencies>
 </project>
