+++ java/src/main/java/com/code972/hebmorph/hspell/HSpellLoader.java
@@ -19,12 +19,7 @@ package com.code972.hebmorph.hspell;
 
 import com.code972.hebmorph.DictionaryLoader;
 import com.code972.hebmorph.MorphData;
-import com.code972.hebmorph.datastructures.DictHebMorph;
 import com.code972.hebmorph.datastructures.DictRadix;
-import org.apache.lucene.analysis.hebrew.HebrewExactAnalyzer;
-import org.apache.lucene.analysis.hebrew.HebrewIndexingAnalyzer;
-import org.apache.lucene.analysis.hebrew.HebrewQueryAnalyzer;
-import org.apache.lucene.analysis.hebrew.HebrewQueryLightAnalyzer;
 
 import java.io.*;
 import java.nio.charset.Charset;
@@ -452,29 +447,5 @@ public final class HSpellLoader {
             num += digit;
         }
         return num;
-    }
-
-    public static HebrewIndexingAnalyzer getHebrewIndexingAnalyzer() throws IOException {
-        DictRadix<MorphData> radix = new HSpellLoader(new File(HSpellLoader.getHspellPath()), true).loadDictionaryFromHSpellData();
-        HashMap<String, Integer> prefs = HSpellLoader.readPrefixesFromFile(HSpellLoader.getHspellPath() + HSpellLoader.PREFIX_NOH);
-        return new HebrewIndexingAnalyzer(new DictHebMorph(radix, prefs));
-    }
-
-    public static HebrewQueryAnalyzer getHebrewQueryAnalyzer() throws IOException {
-        DictRadix<MorphData> radix = new HSpellLoader(new File(HSpellLoader.getHspellPath()), true).loadDictionaryFromHSpellData();
-        HashMap<String, Integer> prefs = HSpellLoader.readPrefixesFromFile(HSpellLoader.getHspellPath() + HSpellLoader.PREFIX_NOH);
-        return new HebrewQueryAnalyzer(new DictHebMorph(radix, prefs));
-    }
-
-    public static HebrewQueryLightAnalyzer getHebrewQueryLightAnalyzer() throws IOException {
-        DictRadix<MorphData> radix = new HSpellLoader(new File(HSpellLoader.getHspellPath()), true).loadDictionaryFromHSpellData();
-        HashMap<String, Integer> prefs = HSpellLoader.readPrefixesFromFile(HSpellLoader.getHspellPath() + HSpellLoader.PREFIX_NOH);
-        return new HebrewQueryLightAnalyzer(new DictHebMorph(radix, prefs));
-    }
-
-    public static HebrewExactAnalyzer getHebrewExactAnalyzer() throws IOException {
-        DictRadix<MorphData> radix = new HSpellLoader(new File(HSpellLoader.getHspellPath()), true).loadDictionaryFromHSpellData();
-        HashMap<String, Integer> prefs = HSpellLoader.readPrefixesFromFile(HSpellLoader.getHspellPath() + HSpellLoader.PREFIX_NOH);
-        return new HebrewExactAnalyzer(new DictHebMorph(radix, prefs));
     }
 }
