+++ pom.xml
@@ -1,14 +1,39 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
-                      http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0                       http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.code972.hebmorph</groupId>
-  <artifactId>hebmorph-java</artifactId>
-  <version>2.0</version>
-  <packaging>pom</packaging>
+  <parent>
+    <groupId>org.sonatype.oss</groupId>
+    <artifactId>oss-parent</artifactId>
+    <version>7</version>
+  </parent>
 
+  <groupId>com.code972.hebmorph</groupId>
+  <artifactId>hebmorph-lucene</artifactId>
+  <version>2.0-SNAPSHOT</version>
+  <packaging>pom</packaging>
+  <name>hebmorph-lucene</name>
+  <description>Enabling Hebrew search through Lucene. Part of the HebMorph project, for making Hebrew properly searchable</description>
+  <url>http://code972.com/</url>
+  <licenses>
+    <license>
+      <name>Affero GPL3</name>
+      <url>http://www.gnu.org/licenses/agpl.html</url>
+      <distribution>repo</distribution>
+      <comments>The GNU Affero General Public License is a free, copyleft license for software and other kinds of works, specifically designed to ensure cooperation with the community in the case of network server software.</comments>
+    </license>
+  </licenses>
+  <developers>
+    <developer>
+      <id>synhershko</id>
+      <name>Itamar Syn-Hershko</name>
+      <url>http://code972.com</url>
+    </developer>
+  </developers>
+  <scm>
+    <url>https://github.com/synhershko/HebMorph</url>
+    <connection>scm:git:ssh://git@github.com/synhershko/HebMorph.git</connection>
+    <developerConnection>scm:git:ssh://git@github.com/synhershko/HebMorph.git</developerConnection>
+  </scm>
   <modules>
     <module>java</module>
   </modules>
